code lint 정리
